+++ socket/src/commonTest/kotlin/mqtt/socket/SimpleSocketTests.kt
@@ -22,14 +22,11 @@ Host: example.com
 Connection: close
 
 """
-        client.write(request)
-        val remotePort = client.remotePort()!!
         val response = client.read().result
         assertTrue { response.contains("200 OK") }
         assertTrue { response.contains("HTTP") }
         assertTrue { response.contains("<html>") }
         client.close()
-        checkPort(remotePort)
     }
 
     @Test
